Refactor code for creating call to static constructor into helper class
