+++ ILCompiler/Compiler/EvaluationStack/CommaEntry.cs
@@ -5,7 +5,7 @@
         public StackEntry Op1 { get; }
         public StackEntry Op2 { get; }
 
-        public CommaEntry(StackEntry op1, StackEntry op2, VarType type) : base(type, type.GetTypeSize())
+        public CommaEntry(StackEntry op1, StackEntry op2) : base(op2.Type)
         {
             Op1 = op1;
             Op2 = op2;
@@ -13,7 +13,7 @@
 
         public override StackEntry Duplicate()
         {
-            return new CommaEntry(Op1.Duplicate(), Op2.Duplicate(), Type);
+            return new CommaEntry(Op1.Duplicate(), Op2.Duplicate());
         }
 
         public override void Accept(IStackEntryVisitor visitor)
