+++ ILCompiler/Compiler/Importer/LoadFieldImporter.cs
@@ -30,7 +30,7 @@ namespace ILCompiler.Compiler.Importer
                 var mangledFieldName = context.NameMangler.GetMangledFieldName(fieldDef);
                 obj = new StaticFieldEntry(mangledFieldName);
 
-                obj = ImportInitClass(fieldDef, context, importer, obj);
+                obj = InitClassHelper.ImportInitClass(fieldDef, context, importer, obj);
             }
             else
             {
@@ -54,24 +54,6 @@ namespace ILCompiler.Compiler.Importer
             importer.PushExpression(node);
 
             return true;
-        }
-
-        private static StackEntry ImportInitClass(FieldDef fieldDef, ImportContext context, IILImporterProxy importer, StackEntry obj)
-        {
-            // Get the static constructor if one exists
-            var declaringType = fieldDef.DeclaringType;
-            var staticConstructorMethod = declaringType.FindStaticConstructor();
-            if (staticConstructorMethod == null)
-            {
-                return obj;
-            }
-
-            // Generate call to static constructor
-            // TODO: NEED TO ENSURE THIS IS ONLY CALLED ONCE THOUGH.
-            // idea is to modify code in static constructor so that at the end of the method it changes the initial code to a RET
-            var targetMethod = context.NameMangler.GetMangledMethodName(staticConstructorMethod);
-            var staticInitCall = new CallEntry(targetMethod, new List<StackEntry>(), VarType.Void, 0);
-            return new CommaEntry(staticInitCall, obj, obj.Type);
         }
 
         private static StackEntry GetStructAddress(ref uint fieldOffset, StackEntry structVal, IILImporterProxy importer)
